Second Commit
Added picture about me and a short paragraph about myself and my travels. Change the background colors on the "About Me"  jumbotron section and "Hecho en Mexico" section.  In addition, added local links to my own site to the approriate section. For example, clicking on the Dev/Mission banner in the carousel would lead the user to the Dev.Mission section on my site.